fix our sound issue
1. sound works now
2. We were arranging our bytes in the wrong order
When we read the two bytes, we shifted the first byte and then appended
the second byte.
This is wrong because it is little endian. We should shift the second
byte, and append the first byte.

Now it is a matter of finding background music and sound effects and
combining them.