update ai
Basically extend raymond's idea ...

AI have two modes of operation: Random and Directed...
Random, it has no clue where to hit, so it randomly generate
coordinates. When it hits something, then it inserts the 4 adjacent
coordinates into a Stack, given if they have not been visited. If this
Stack of potential targets is not empty, then the AI operates in
Directed mode. Instead of generating random targets, it just pops from
the stack.