fix a bug regarding HP in single player mode
In single player, the player has HP.
HP is added when a ship is added onto the board.
However, when a ship is taken off the board (long press), the hp doesn't
get subtracted.
The fix here is to add a setter for the HP, then set the value of the HP
when a ship gets removed.